Styling: Added Margin on Current & Upcoming Events
Also sneaking in a new bug_report SVG that was cooked up briefly, trying
to make the bug image used seem a little less blurred.

I didn't actually make any new class for this modification, but just
removed the zero-margin class that was applied to the header (after
permission). This might eventually get reversed again, but for now, the
title needs space.