Explicitly set rails config time to UTC
It's worthwhile to explicitly state that the server is using UTC time.
All user input should be converted to UTC ASAP, and all time data about
to be shown to the user should be converted to their requested timezone.
This way stuff like daylight savings shouldn't need explicit conditions -
it will be handled automatically.