Added group privacy to db and switched constants from strings to enums
Enums are backed by integers, so 'role' in users_group was switched
to an integer as well. Really I just wanted to explicitly declare
our constants and scope them for rails, but ActiveSupport just happens
to back these with integers in the database (probably because it's more
efficient for large tables).