Added group views
There's a lot of overlapping code between the user and group views
right now, so I made sure their controllers dealt with tabs/pages similarly.
Namely, tab page logic should probably be in the controller, and if not a
controller, then in a presenter/decorator.

As can be seen, not all of the group views are done yet. This is a partial
commit.

Lastly, I apologize, but RubyMine is fighting me on using tabs. There are
plenty of spots where the project mixes spaces and tabs anyhow, so I suppose
it's a concern that should be brought up anyway.