Updated user and group models
Paperclip settings are now stored in application.rb. In addition,
the Group privacies and User-Group role enums have been given a
more sensible name.

A where clause was also added to both relationship models to account
for the fact that group_invitations are being stored in the users_groups
table. I believe the more ideal model would be to keep a separate
GroupInvitations table in the schema, but this also works.