Removed XSS vulnerability
This commit adds escapeHtml to dynamically generated dom elements to
prevent things like <script>alert(1);</script> from being executed.

In addition, this changes the default value of category names stored in
the database from <i>Untitled</i> to an empty string. The reason behind this
is that <i>Untitled</i> is a legitimate category name, whereas an empty
string is not. If category names are allowed to contain HTML, then how
the default category name is styled needs to be rethought.