Added initial Maps iframe
This is pretty sluggish at the moment, as it will reload on each overlay load.

Also, this uses a key originally intended for Northcott, so I own it at the moment. It's pretty simple for us to obtain a key though - you basically click a button in the Maps API page. We'll talk about managing keys and whether or not it is wise to even do this branch soon, though.