Fully updated to Rails 5.2 (except active-storage)
There are a bunch of small changes like `belongs_to` having to explicitly
state if a resource is optional.

A couple things (like using `image_tag`) and storage.yml are for the eventual move
to active-storage.

Lastly, bootsnap was added. Launching rails the first time will be as slow as usual,
but subsequent relaunches start twice as fast.