Excluded existing !important from scss linter
The only case Carpe actually needs `!important` is for overriding
the token-input-input-token because for some god-forsaken reason they
apply the width as a style attribute using Javascript. Every other
use-case can and should be resolved through proper import and
rule declaration ordering. Since Id styling selectors are already
being used in several locations, removing the use of important is
too difficult and fragile for the scope of this commit's branch.